+++ Anypic/PAPMessagingViewController.m
@@ -434,7 +434,7 @@ static inline UIViewAnimationOptions animationOptionsWithCurve(UIViewAnimationCu
         
         cell.userInteractionEnabled = NO;
         
-        if (fabsf(timeInterval) > (12 * 60 * 60)) {
+        if (fabs(timeInterval) > (12 * 60 * 60)) {
             cell.timeStampLabel.text = timestamp;
         } else {
             cell.timeStampLabel.text = [dateFormat stringFromDate:updatedDate];
